fix(errorprone): widen to long before integer multiplication / subtraction (IntLongMath, NarrowCalculation)
ErrorProne flags int arithmetic that gets implicitly widened to long
because the int expression itself can overflow before the cast. The
flagged sites here are unlikely to overflow in practice but cheap to
fix and now make the widening explicit:

- FileComparator: permission diff into long.
- ZipFile.centralHeaderLen: 46L + ints.
- ZipEntry.javaToDosTime: cast year-1980 to long before <<25.
- NrgParser.tracksEnd: cast toDword() result to long before -1.
- SizeFormat: cast nKB/nMB to long before multiplying by KB_1/MB_1
  inside the size-N*unit subtraction.

+++ mucommander-commons-file/src/main/java/com/mucommander/commons/file/util/FileComparator.java
@@ -256,7 +256,7 @@ public class FileComparator implements Comparator<AbstractFile> {
             diff = f1.getDate()-f2.getDate();
             break;
         case PERMISSIONS:
-            diff = f1.getPermissions().getIntValue() - f2.getPermissions().getIntValue();
+            diff = (long) f1.getPermissions().getIntValue() - f2.getPermissions().getIntValue();
             break;
         case EXTENSION:
             diff = StringUtils.compareIgnoreCase(f1.getExtension(), f2.getExtension());
